requirements.txt
# Core LLM stack
openai
langchain-openai
langchain-community
langchain-pinecone

# Vector database
pinecone

# Web / API / UI
fastapi
uvicorn[standard]
streamlit
requests

# PDF parsing
PyPDF2

# Utilities
python-dotenv
tqdm
pydantic